9.2.RC1-GenetICS-PLUS with 3.0.8 kernel
Kernel switched to:
https://github.com/androidarmv6/android_kernel_lge_msm7x27-3.0.x/commits/cm-10.1

Technically, this kernel supports all ics and jellybean roms if rom built with 3.0 kernel headers.
This rom is already built with 3.0.8 headers and all 3.0.8 releases from github.com/DJNoXD will support it.

Anyway, 9.2.RC1 is last ROM from me. If DJNoXD releases a stable kernel, this rom also will be stable.
And you can think RC1 means STABLE.

As a result, this ROM's lifespan is finished. But, KERNEL developing will be alive if DJNoXD supports it.

My last word is follow here:
http://forum.xda-developers.com/showthread.php?t=2144790
AND THANK HIM A LOT.....................................